# store-gateway: remove internal LabelValues