--- a/pkg/storegateway/indexheader/lazy_binary_reader.go
+++ b/pkg/storegateway/indexheader/lazy_binary_reader.go
@@ -176,19 +176,6 @@ func (r *LazyBinaryReader) LookupSymbol(o uint32) (string, error) {
 	return r.reader.LookupSymbol(o)
 }
 
-// LabelValues implements Reader.
-func (r *LazyBinaryReader) LabelValues(name string, prefix string, filter func(string) bool) ([]string, error) {
-	r.readerMx.RLock()
-	defer r.readerMx.RUnlock()
-
-	if err := r.load(); err != nil {
-		return nil, err
-	}
-
-	r.usedAt.Store(time.Now().UnixNano())
-	return r.reader.LabelValues(name, prefix, filter)
-}
-
 // LabelValuesOffsets implements Reader.
 func (r *LazyBinaryReader) LabelValuesOffsets(name string, prefix string, filter func(string) bool) ([]streamindex.PostingListOffset, error) {
 	r.readerMx.RLock()