#### What this PR does

As part of #4593 I replaced the usages of `LabelValues() []string` with `LabelValuesOffsets() []PostingListOffset`

This PR is in internal clean-up, which removes the implementation of LabelValues from the postings tables and index header readers.
It also replaces a few tests and benchmarks to use the new implementation.


#### Which issue(s) this PR fixes or relates to

Fixes #<issue number>

#### Checklist

- [x] Tests updated
- [n/a] Documentation added
- [n/a] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`