@pracucci 2023-05-05
> I went ahead and further simplified PostingListOffset in https://github.com/grafana/mimir/commit/9d309d3b69183c379b79baba66188f08a07b5bc5. Let me know if you want to take another look

@dimitarvdimitrov could you run a benchmark touching `LabelValuesOffsets()`, please? I want to make sure the new data structure `pEntry` doesn't introduce any performance regression.