@dimitarvdimitrov 2023-05-05
I ran the benchmarks and there was no regression. There is a significant decrease in CPU, but I think that was because of noise on my machine.

There is the overhead of having a larger data structure for label values (struct vs string). But no regression except that.

<details>
<summary>Details</summary>

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/storegateway/indexheader
                                                                               │ Rename-tests-ba06a1ead.txt │ Further-simplify-PostingListOffset-7e8e9e717.txt │
                                                                               │           sec/op           │          sec/op            vs base               │
LabelValuesOffsetsIndexV2_WithPrefix/test_label_5/prefix=''-10                                 11.95µ ± ∞ ¹                11.81µ ± ∞ ¹        ~ (p=0.686 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_5/prefix='value_'-10                           12.27µ ± ∞ ¹                11.73µ ± ∞ ¹   -4.40% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_5/prefix='value_000'-10                        11.71µ ± ∞ ¹                11.35µ ± ∞ ¹   -3.09% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_5/prefix='value_001'-10                        12.03µ ± ∞ ¹                11.46µ ± ∞ ¹        ~ (p=0.057 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_5/prefix='value_004'-10                        11.88µ ± ∞ ¹                11.97µ ± ∞ ¹        ~ (p=0.686 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_5/prefix='value_005'-10                        563.9n ± ∞ ¹                560.0n ± ∞ ¹        ~ (p=0.114 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_29/prefix=''-10                                13.23µ ± ∞ ¹                13.77µ ± ∞ ¹        ~ (p=0.200 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_29/prefix='value_'-10                          13.26µ ± ∞ ¹                13.49µ ± ∞ ¹   +1.73% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_29/prefix='value_000'-10                       11.77µ ± ∞ ¹                11.73µ ± ∞ ¹        ~ (p=0.886 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_29/prefix='value_001'-10                       11.70µ ± ∞ ¹                11.59µ ± ∞ ¹        ~ (p=0.686 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_29/prefix='value_028'-10                       12.05µ ± ∞ ¹                11.46µ ± ∞ ¹   -4.88% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_29/prefix='value_029'-10                       567.4n ± ∞ ¹                557.5n ± ∞ ¹   -1.74% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_8/prefix=''-10                                 11.88µ ± ∞ ¹                11.81µ ± ∞ ¹        ~ (p=0.486 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_8/prefix='value_'-10                           11.83µ ± ∞ ¹                11.91µ ± ∞ ¹   +0.73% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_8/prefix='value_000'-10                        11.41µ ± ∞ ¹                11.38µ ± ∞ ¹        ~ (p=0.886 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_8/prefix='value_001'-10                        11.80µ ± ∞ ¹                11.47µ ± ∞ ¹        ~ (p=0.200 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_8/prefix='value_007'-10                        11.37µ ± ∞ ¹                11.38µ ± ∞ ¹        ~ (p=0.886 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_8/prefix='value_008'-10                        574.8n ± ∞ ¹                558.7n ± ∞ ¹   -2.80% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_19/prefix=''-10                                13.76µ ± ∞ ¹                12.72µ ± ∞ ¹   -7.56% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_19/prefix='value_'-10                          14.15µ ± ∞ ¹                12.78µ ± ∞ ¹   -9.68% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_19/prefix='value_000'-10                       13.12µ ± ∞ ¹                11.48µ ± ∞ ¹  -12.45% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_19/prefix='value_001'-10                       12.29µ ± ∞ ¹                11.53µ ± ∞ ¹   -6.23% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_19/prefix='value_018'-10                       11.85µ ± ∞ ¹                11.49µ ± ∞ ¹   -3.02% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_19/prefix='value_019'-10                       621.3n ± ∞ ¹                555.5n ± ∞ ¹  -10.58% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_23/prefix=''-10                                14.90µ ± ∞ ¹                12.87µ ± ∞ ¹  -13.62% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_23/prefix='value_'-10                          13.42µ ± ∞ ¹                12.98µ ± ∞ ¹   -3.21% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_23/prefix='value_000'-10                       11.89µ ± ∞ ¹                11.70µ ± ∞ ¹        ~ (p=0.114 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_23/prefix='value_001'-10                       11.76µ ± ∞ ¹                11.71µ ± ∞ ¹        ~ (p=0.343 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_23/prefix='value_022'-10                       11.64µ ± ∞ ¹                11.52µ ± ∞ ¹        ~ (p=0.114 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_23/prefix='value_023'-10                       580.8n ± ∞ ¹                558.5n ± ∞ ¹   -3.82% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_31/prefix=''-10                                13.80µ ± ∞ ¹                13.11µ ± ∞ ¹   -5.00% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_31/prefix='value_'-10                          13.77µ ± ∞ ¹                13.39µ ± ∞ ¹   -2.79% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_31/prefix='value_000'-10                       11.66µ ± ∞ ¹                11.46µ ± ∞ ¹   -1.67% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_31/prefix='value_001'-10                       12.56µ ± ∞ ¹                11.42µ ± ∞ ¹   -9.09% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_31/prefix='value_030'-10                       12.35µ ± ∞ ¹                11.40µ ± ∞ ¹        ~ (p=0.114 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_31/prefix='value_031'-10                       588.1n ± ∞ ¹                557.2n ± ∞ ¹   -5.25% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_4/prefix=''-10                                 12.10µ ± ∞ ¹                11.58µ ± ∞ ¹   -4.26% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_4/prefix='value_'-10                           12.17µ ± ∞ ¹                11.67µ ± ∞ ¹   -4.12% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_4/prefix='value_000'-10                        11.65µ ± ∞ ¹                11.49µ ± ∞ ¹        ~ (p=0.200 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_4/prefix='value_001'-10                        12.76µ ± ∞ ¹                11.53µ ± ∞ ¹   -9.66% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_4/prefix='value_003'-10                        12.34µ ± ∞ ¹                11.48µ ± ∞ ¹        ~ (p=0.343 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_4/prefix='value_004'-10                        574.0n ± ∞ ¹                558.5n ± ∞ ¹   -2.72% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_6/prefix=''-10                                 12.82µ ± ∞ ¹                11.92µ ± ∞ ¹   -7.01% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_6/prefix='value_'-10                           12.12µ ± ∞ ¹                11.91µ ± ∞ ¹        ~ (p=0.200 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_6/prefix='value_000'-10                        13.17µ ± ∞ ¹                11.50µ ± ∞ ¹  -12.67% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_6/prefix='value_001'-10                        12.52µ ± ∞ ¹                11.53µ ± ∞ ¹   -7.92% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_6/prefix='value_005'-10                        11.91µ ± ∞ ¹                11.48µ ± ∞ ¹   -3.56% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_6/prefix='value_006'-10                        614.8n ± ∞ ¹                557.3n ± ∞ ¹   -9.35% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_15/prefix=''-10                                12.72µ ± ∞ ¹                12.38µ ± ∞ ¹   -2.66% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_15/prefix='value_'-10                          14.28µ ± ∞ ¹                12.40µ ± ∞ ¹  -13.17% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_15/prefix='value_000'-10                       13.73µ ± ∞ ¹                11.40µ ± ∞ ¹  -16.95% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_15/prefix='value_001'-10                       12.95µ ± ∞ ¹                11.45µ ± ∞ ¹  -11.60% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_15/prefix='value_014'-10                       12.57µ ± ∞ ¹                11.50µ ± ∞ ¹   -8.47% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_15/prefix='value_015'-10                       580.5n ± ∞ ¹                556.2n ± ∞ ¹   -4.17% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_20/prefix=''-10                                15.25µ ± ∞ ¹                12.76µ ± ∞ ¹  -16.37% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_20/prefix='value_'-10                          14.51µ ± ∞ ¹                12.82µ ± ∞ ¹  -11.69% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_20/prefix='value_000'-10                       12.86µ ± ∞ ¹                11.51µ ± ∞ ¹  -10.50% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_20/prefix='value_001'-10                       12.27µ ± ∞ ¹                11.55µ ± ∞ ¹   -5.91% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_20/prefix='value_019'-10                       11.93µ ± ∞ ¹                11.39µ ± ∞ ¹   -4.52% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_20/prefix='value_020'-10                       575.3n ± ∞ ¹                557.6n ± ∞ ¹   -3.09% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_2/prefix=''-10                                 12.53µ ± ∞ ¹                11.64µ ± ∞ ¹   -7.10% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_2/prefix='value_'-10                           14.13µ ± ∞ ¹                11.64µ ± ∞ ¹  -17.63% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_2/prefix='value_000'-10                        14.04µ ± ∞ ¹                11.51µ ± ∞ ¹  -18.00% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_2/prefix='value_001'-10                        12.73µ ± ∞ ¹                11.46µ ± ∞ ¹  -10.01% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_2/prefix='value_001'#01-10                     13.01µ ± ∞ ¹                11.68µ ± ∞ ¹  -10.25% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_2/prefix='value_002'-10                        632.2n ± ∞ ¹                562.9n ± ∞ ¹  -10.98% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_12/prefix=''-10                                14.25µ ± ∞ ¹                12.15µ ± ∞ ¹  -14.75% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_12/prefix='value_'-10                          14.93µ ± ∞ ¹                12.15µ ± ∞ ¹  -18.64% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_12/prefix='value_000'-10                       12.35µ ± ∞ ¹                11.40µ ± ∞ ¹   -7.68% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_12/prefix='value_001'-10                       12.23µ ± ∞ ¹                11.50µ ± ∞ ¹   -5.98% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_12/prefix='value_011'-10                       12.66µ ± ∞ ¹                11.70µ ± ∞ ¹        ~ (p=0.114 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_12/prefix='value_012'-10                       581.8n ± ∞ ¹                560.8n ± ∞ ¹   -3.62% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_14/prefix=''-10                                13.12µ ± ∞ ¹                12.30µ ± ∞ ¹   -6.25% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_14/prefix='value_'-10                          12.93µ ± ∞ ¹                12.39µ ± ∞ ¹   -4.15% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_14/prefix='value_000'-10                       11.96µ ± ∞ ¹                11.52µ ± ∞ ¹   -3.67% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_14/prefix='value_001'-10                       12.11µ ± ∞ ¹                11.53µ ± ∞ ¹   -4.77% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_14/prefix='value_013'-10                       12.89µ ± ∞ ¹                11.47µ ± ∞ ¹  -11.00% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_14/prefix='value_014'-10                       697.7n ± ∞ ¹                555.4n ± ∞ ¹  -20.40% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_16/prefix=''-10                                14.92µ ± ∞ ¹                12.58µ ± ∞ ¹  -15.71% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_16/prefix='value_'-10                          13.83µ ± ∞ ¹                12.91µ ± ∞ ¹        ~ (p=0.057 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_16/prefix='value_000'-10                       12.56µ ± ∞ ¹                11.66µ ± ∞ ¹   -7.15% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_16/prefix='value_001'-10                       12.19µ ± ∞ ¹                11.66µ ± ∞ ¹        ~ (p=0.114 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_16/prefix='value_015'-10                       12.30µ ± ∞ ¹                11.56µ ± ∞ ¹   -6.02% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_16/prefix='value_016'-10                       585.2n ± ∞ ¹                556.9n ± ∞ ¹   -4.84% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_27/prefix=''-10                                13.77µ ± ∞ ¹                12.90µ ± ∞ ¹   -6.31% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_27/prefix='value_'-10                          14.01µ ± ∞ ¹                13.02µ ± ∞ ¹   -7.05% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_27/prefix='value_000'-10                       11.99µ ± ∞ ¹                11.33µ ± ∞ ¹   -5.54% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_27/prefix='value_001'-10                       12.23µ ± ∞ ¹                11.42µ ± ∞ ¹   -6.55% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_27/prefix='value_026'-10                       11.91µ ± ∞ ¹                11.40µ ± ∞ ¹   -4.23% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_27/prefix='value_027'-10                       586.3n ± ∞ ¹                554.3n ± ∞ ¹   -5.46% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_9/prefix=''-10                                 12.81µ ± ∞ ¹                11.89µ ± ∞ ¹   -7.16% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_9/prefix='value_'-10                           12.57µ ± ∞ ¹                12.03µ ± ∞ ¹        ~ (p=0.057 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_9/prefix='value_000'-10                        11.97µ ± ∞ ¹                11.41µ ± ∞ ¹   -4.69% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_9/prefix='value_001'-10                        11.97µ ± ∞ ¹                11.37µ ± ∞ ¹   -5.01% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_9/prefix='value_008'-10                        11.90µ ± ∞ ¹                11.30µ ± ∞ ¹   -5.09% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_9/prefix='value_009'-10                        598.5n ± ∞ ¹                556.8n ± ∞ ¹   -6.98% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_10/prefix=''-10                                12.67µ ± ∞ ¹                11.95µ ± ∞ ¹   -5.71% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_10/prefix='value_'-10                          12.88µ ± ∞ ¹                12.11µ ± ∞ ¹   -5.94% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_10/prefix='value_000'-10                       11.99µ ± ∞ ¹                11.44µ ± ∞ ¹   -4.56% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_10/prefix='value_001'-10                       12.19µ ± ∞ ¹                11.48µ ± ∞ ¹   -5.82% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_10/prefix='value_009'-10                       12.07µ ± ∞ ¹                11.42µ ± ∞ ¹   -5.40% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_10/prefix='value_010'-10                       582.0n ± ∞ ¹                553.5n ± ∞ ¹   -4.89% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_22/prefix=''-10                                13.38µ ± ∞ ¹                12.70µ ± ∞ ¹   -5.05% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_22/prefix='value_'-10                          13.66µ ± ∞ ¹                12.81µ ± ∞ ¹   -6.24% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_22/prefix='value_000'-10                       12.40µ ± ∞ ¹                11.55µ ± ∞ ¹   -6.85% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_22/prefix='value_001'-10                       12.26µ ± ∞ ¹                11.94µ ± ∞ ¹   -2.67% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_22/prefix='value_021'-10                       12.07µ ± ∞ ¹                11.45µ ± ∞ ¹   -5.12% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_22/prefix='value_022'-10                       580.0n ± ∞ ¹                554.2n ± ∞ ¹   -4.44% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_24/prefix=''-10                                13.51µ ± ∞ ¹                12.84µ ± ∞ ¹   -5.02% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_24/prefix='value_'-10                          13.64µ ± ∞ ¹                12.82µ ± ∞ ¹   -6.02% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_24/prefix='value_000'-10                       12.07µ ± ∞ ¹                11.37µ ± ∞ ¹   -5.78% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_24/prefix='value_001'-10                       12.21µ ± ∞ ¹                11.46µ ± ∞ ¹   -6.14% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_24/prefix='value_023'-10                       12.02µ ± ∞ ¹                11.49µ ± ∞ ¹   -4.41% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_24/prefix='value_024'-10                       583.9n ± ∞ ¹                561.9n ± ∞ ¹   -3.76% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_25/prefix=''-10                                13.68µ ± ∞ ¹                13.00µ ± ∞ ¹   -5.00% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_25/prefix='value_'-10                          13.75µ ± ∞ ¹                13.13µ ± ∞ ¹   -4.49% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_25/prefix='value_000'-10                       12.10µ ± ∞ ¹                11.55µ ± ∞ ¹   -4.48% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_25/prefix='value_001'-10                       12.13µ ± ∞ ¹                11.48µ ± ∞ ¹   -5.30% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_25/prefix='value_024'-10                       12.06µ ± ∞ ¹                11.36µ ± ∞ ¹   -5.76% (p=0.029 n=4)
LabelValuesOffsetsIndexV2_WithPrefix/test_label_25/prefix='value_025'-10                       581.3n ± ∞ ¹                556.8n ± ∞ ¹   -4.22% (p=0.029 n=4)
…